Support reception of 64-bit data-sequence-numbers
We introduce a circular seq-number array. This array contains the
32 high-order bits of the sequence-number that may currently be
considered.

We have two arrays, snd_high_order and rcv_high_order in the mpcb. And a
bit to indicate which is the "current" one.

For the sending-side, the current-one is the high-order of the
snd_nxt. Thus, the second value of snd_high_order is always one less
than the current-one, as retransmissions are always < snd_nxt.

For the receiving-side, the current-one is the high-order of the
rcv_nxt. The second value of rcv_high_order is always one more than the
current-one, as we expect the next segment to >= rcv_nxt.
If it is in reality smaller, our high-order guess may be wrong. This
doesn't matter, as it will then be dropped because the DSS-csum will
fail. As the peer only sent 32-bit seq-number we can only guess it. If
he sends a 64-bit seq-number we don't have to guess and can drop the
segment due to out-of-window-ness.

+++ include/linux/tcp.h
@@ -535,8 +535,8 @@ struct tcp_sock {
 	} bw_est;
 	u32	cur_bw_est;
 
-	/* per subflow data, for tcp_recvmsg */
-	u32	map_data_seq; /* Those three fields record the current mapping */
+	 /* Those three fields record the current mapping */
+	u64	map_data_seq;
 	u32	map_subseq;
 	u16	map_data_len;
 
@@ -560,7 +560,7 @@ struct tcp_sock {
 	int		path_index;
 	struct tcp_sock	*next;		/* Next subflow socket */
 	__u32		mptcp_loc_random_number;
-	uint8_t		slave_sk:1,
+	u8		slave_sk:1,
 			fully_established:1,
 			attached:1,
 			csum_error:1,
